Refactoring of event handler
